+++ go.mod
@@ -1,0 +1,3 @@
+module mcronalds
+
+go 1.21.1
