+++ produto/produto.go
@@ -1,0 +1,9 @@
+package produto
+
+type Produto struct {
+	ID        int
+	Nome      string
+	Descricao string
+	Preco     float64
+}
+
